9commits

+++ database/migrations/2021_02_21_030456_create_followers_table.php
@@ -14,8 +14,8 @@ class CreateFollowersTable extends Migration
     public function up()
     {
         Schema::create('followers', function (Blueprint $table) {
-            $table->unsignedInteger('following_id')->comment('フォローしているユーザID');
-            $table->unsignedInteger('followed_id')->comment('フォローされているユーザID');
+            $table->unsignedBigInteger('following_id')->comment('フォローしているユーザID');
+            $table->unsignedBigInteger('followed_id')->comment('フォローされているユーザID');
 
             $table->index('following_id');
             $table->index('followed_id');
